// src/app/shared/enums/lookup.ts
export enum Lookup {
  None = -1,
  Building = 1,
  Floor = 2,
  Asset_Type = 3,
  Asset_Replace = 4,
  Technical_Department_ = 5,
  Generic = 6,
  Currency = 7,
  Parent_Childs = 8,
  Technical_guidance_books = 9,
  Tax = 10,
  City = 11,
  Collection_Period = 12,
  Cost_Of_Customer = 13,
  Cust_Category = 14,
  Cust_Group = 15,
  Cust_Type = 16,
  Sales_SubRep = 17,
  Country_Of_Origin = 18,
  PPM_Frequency = 19,
  Warranty_Period = 20,
  Asset_Risk_Level = 21,
  Asset_Business_Critical_Level = 22,
  Asset_Categories = 23,
  AssetClassifications = 24,
  Asset_Function_Class = 25,
  Asset_Modility = 26,
  Sub_Modility = 27,
  AssetType = 28,
  Contract_Type = 29,
  Contract_status = 30,
  MType_Of_Contract = 31,
  Time_Period = 32,
  Assigned_To = 33,
  Type_Of_Service = 34,
  Supplier_Status = 35,
  Person_Role = 36,
  Rooms = 37,
  Acquisition_Method = 100,
  Budget_Year = 101,
  Commissioning_Status = 102,
  Months = 103,
  Ppm_time_period = 104,
  User_Position = 105,
  userQualifications = 106,
  userGroups = 107,
  ServiceTeam = 108,
  Operator_Date = 200,
  MAINTANAN_CECONTRACT = 201,
  Asset_Classifications = 202,
  Asset_Status = 203,
  StockClass = 208,
  SpareInfoGrouping = 209,
  PartType = 210,
  DefectType = 601,
  Priority = 602,
  requestedThrough = 603,
  TypeOfRequest = 604,
  SparePartTransactionStatus = 211,
  SparePartReserved = 212,
  PartDeliveryWorkFlowApproval = 213,
  PartDeliveryStatusOfPermission = 214,
  SurveyStatus = 215,
  AssetTypeWO = 500,
  TypeofWO = 501,
  callReview = 404,
  MachineStatus = 406,
  BackUpStatus = 459,
  NextBackUpStatus = 458,
  CallslastSituationWO = 502,
  StatusWO = 503,
  RepairLocation = 504,
  Reasons = 505,
  First_Action = 700,
  GazType = 606,
  SizeCylinder = 607,
  CylinderSize = 608,
  StatusGazRefill = 609,
  SupplierCodeTypes = 407,
  AssetNDCodeTypes = 408,
  ModelDefinitionCodeTypes = 405,
  QuotationType = 515,
  QuotationStatus = 516,
  QuotationOfferType = 517,
  QuotationCallLastSituation = 518,
  POAssetStatut = 519,
  POCallLastSituation = 520,
  POStatus = 521,
  POReasons = 522,
  POComponentToBeFixed = 523,
  POPartStatus = 524,
  Voltage = 525,
  Frequency = 526,
  PowerSupplyType = 527,
  RegistryFiles = 528,
  InspectionTest = 529,
  InspectionAlert = 530,
  DeliveryType = 531,
  DeliveryStatus = 532,
  AssetDeliverySites = 533,
  EndUserAcceptanceStatus = 534,

  Positions = 410,
  Rates = 411,
  ServiceTeams = 412,
  TRAFRequestType = 416,
  AssignOrigin = 418,
  TrafAssessClassification = 450,
  TrafAssessRecommendation = 451,
  FirstActionConfirmation = 702,
  NoOfCall = 703,
  SurveyStatus1 = 420,
  LookupValueQuestion = 421,
  TypeOfSurvey = 422,
  Shift_Type = 40,
  TrafOfferStatus = 455,
  TrafStatusOfDR = 452,
  TrafStatusOfRequester = 453,
  TrafTypeOfOffer = 454,
  FindingType = 42,
  SparePartStatus = 456,
  DemoStatus = 457,
  InternalAuditFix = 800,
  InternalAuditStatus = 41,
  Safety = 480,
  JobStatus = 1200,
  JobPriority = 1201,
  Source = 1202,
  CostCode = 1203,
  ProblemDescription = 1204,
  AssetAvailability = 1205,
  DemoRequestTypye = 1206,
  InstructionDescription = 1207,
  InstructionText = 1208,
  Qty = 1209,
  InstructionDescriptionCategory = 1210,
}